files populated inside core
