+++ src/archemist/core/persistence/object_factory.py
@@ -5,6 +5,8 @@ if TYPE_CHECKING:
     from archemist.core.state.batch import Batch
     from archemist.core.state.robot import Robot, RobotOpDescriptor, RobotModel
     from archemist.core.state.station import Station, StationOpDescriptor, StationModel
+    from archemist.core.state.optimisation_state import OptimizationState
+    from archemist.core.models.optimisation_model import OptimisationModel
 
 import importlib
 import pkgutil
@@ -144,3 +146,16 @@ class StationFactory:
             handler_module = importlib.import_module(station_module)
             cls = getattr(handler_module, handler_type)
         return cls(station)
+
+class OptimizationFactory:
+    @staticmethod
+    def create_from_dict(config_dict: Dict) -> OptimizationState:
+        from archemist.core.state.optimisation_state import OptimizationState
+        return OptimizationState.from_dict(config_dict)
+
+    @staticmethod
+    def create_from_model(optimization_model: OptimisationModel) -> OptimizationState:
+        module = importlib.import_module(optimization_model._module)
+        cls = getattr(module, optimization_model._type)
+        return cls(optimization_model)
+
