+++ tests/optimisation_test_workflow/recipes/algae_bot_recipe_1.yaml
@@ -12,9 +12,9 @@ process:
     operation:
       properties:
         dispense_info:
-          dye_A: [1.0, 0.1, 0.2, 0.82, 0.51, 0.17]
-          dye_B: [0.09, 0.49, 0.78, 0.53, 0.22, 0.93]
-          water: [1.02, 13.49, 44.59, 2.28, 22.92, 17.65]
+          dye_A: [0.59, 0.24, 0.71, 0.01, 0.13, 0.5]
+          dye_B: [0.53, 0.02, 0.19, 0.81, 0.52, 0.23]
+          water: [11.16, 23.09, 27.79, 28.93, 17.35, 3.64]
       type: CSCSVJobOpDescriptor
     type: ChemSpeedFlexStation
   transitions: {on_fail: end_state, on_success: analyse}
