+++ tests/optimisation_test_workflow/recipes/algae_bot_recipe_2.yaml
@@ -12,9 +12,9 @@ process:
     operation:
       properties:
         dispense_info:
-          dye_A: [0.51, 0.97, 0.14, 0.26, 0.27, 0.17]
-          dye_B: [0.4, 0.38, 0.91, 0.33, 0.19, 0.21]
-          water: [18.53, 38.13, 15.33, 47.8, 4.17, 18.09]
+          dye_A: [0.07, 0.78, 0.4, 0.79, 0.97, 0.65]
+          dye_B: [0.92, 0.5, 0.56, 0.45, 0.71, 0.58]
+          water: [44.57, 48.92, 6.61, 19.03, 22.66, 39.7]
       type: CSCSVJobOpDescriptor
     type: ChemSpeedFlexStation
   transitions: {on_fail: end_state, on_success: analyse}
