+++ tests/optimisation_test_workflow/recipes/algae_bot_recipe_3.yaml
@@ -12,9 +12,9 @@ process:
     operation:
       properties:
         dispense_info:
-          dye_A: [0.86, 0.96, 0.03, 0.68, 0.92, 0.81]
-          dye_B: [0.79, 0.75, 0.18, 0.88, 0.1, 0.26]
-          water: [49.76, 39.52, 32.98, 20.0, 4.99, 19.95]
+          dye_A: [0.74, 0.64, 0.49, 0.74, 0.89, 0.07]
+          dye_B: [1.0, 0.27, 0.09, 0.28, 0.94, 0.54]
+          water: [32.71, 38.77, 39.67, 27.34, 3.04, 5.87]
       type: CSCSVJobOpDescriptor
     type: ChemSpeedFlexStation
   transitions: {on_fail: end_state, on_success: analyse}
