+++ tests/optimisation_test_workflow/recipes/template/algae_bot_recipe_template.yaml
